refac: Move all functionality out of AppViewModel
