+++ app/src/main/java/com/mohandass/botforge/chat/ui/components/header/top/DefaultDropdownMenu.kt
@@ -107,7 +107,7 @@ fun DefaultDropdownMenu(viewModel: AppViewModel) {
         )
         DropdownMenuItem(
             onClick = {
-                viewModel.signOut {
+                viewModel.chat.signOut {
                     viewModel.navController.navigate(AppRoutes.Landing.route)
                 }
                 displayOptionsMenu = false
