+++ app/src/main/java/com/mohandass/botforge/settings/viewmodel/SettingsViewModel.kt
@@ -13,6 +13,7 @@ import com.mohandass.botforge.common.services.Analytics
 import com.mohandass.botforge.common.services.snackbar.SnackbarManager
 import com.mohandass.botforge.common.services.snackbar.SnackbarMessage
 import com.mohandass.botforge.common.services.Logger
+import com.mohandass.botforge.common.services.snackbar.SnackbarMessage.Companion.toSnackbarMessageWithAction
 import com.mohandass.botforge.settings.model.PreferredHeader
 import com.mohandass.botforge.settings.model.PreferredTheme
 import com.mohandass.botforge.settings.service.PreferencesDataStore
@@ -154,6 +155,30 @@ class SettingsViewModel @Inject constructor(
                         e.message ?: "Error linking account"
                     )
                 )
+            }
+        }
+    }
+
+    fun signOut(onSuccess: () -> Unit) {
+        viewModelScope.launch {
+            accountService.signOut()
+            onSuccess()
+        }
+    }
+
+    fun deleteAccount(onSuccess: () -> Unit) {
+        viewModelScope.launch {
+            try {
+                accountService.deleteAccount()
+                SnackbarManager.showMessage(R.string.delete_account_success)
+                onSuccess()
+            } catch (e: Exception) {
+                e.toSnackbarMessageWithAction(R.string.sign_out) {
+                    signOut(onSuccess)
+                }.let { message ->
+                    SnackbarManager.showMessage(message)
+                }
+
             }
         }
     }
